Core data implemented into the activity graph. Data is saving from the Add activity tab, only thing is conditioning the arrayed data may be slightly harder.
TO DO:
Get saved data onto the graphs.
Get the weight section saving and then displaying.

+++ Fitness Tracker/AddActivity.h
@@ -21,7 +21,7 @@
 @property (weak, nonatomic) IBOutlet UITextField *HourTextField;
 @property (weak, nonatomic) IBOutlet UITextField *MinuteTextField;
 
-@property NSInteger TimeWorkedOut;
+@property CGFloat TimeWorkedOut;
 
 - (IBAction)SaveButtonPressed:(id)sender;
 - (IBAction)CancelButtonPressed:(id)sender;
